# Game-Score-Manager/game-score-v1.2/backend/src/module/users/dto/role.dto.ts
import { IsEnum } from 'class-validator';
import { Role } from '@prisma/client';

export class ChangeRoleDto {
  @IsEnum(Role) // Validamos que el role esté dentro de los valores válidos del enum Role
  role: Role;
}