# Game-Score-Manager/game-score-v1.2/backend/src/module/users/dto/user.dto.ts
import { Role } from '@prisma/client';
import { IsOptional, IsString, MinLength } from 'class-validator';

export class UpdateUserDto {
  @IsOptional()
  @IsString()
  username: string;

  @IsOptional()
  @IsString()
  @MinLength(8)
  password: string;

  @IsOptional()
  @IsString()
  profileImage?: string;

  @IsOptional()
  @IsString()
  role: Role;
}